+++ libs/config/src/com/biffweb/config.clj
@@ -86,7 +86,7 @@
 
 (defn module
   "On startup, parses config.edn and merges into ctx. Also sets system
-   properties. Include :biff.config/component in your components.
+   properties. Module ID is :biff.config/module.
 
    Loads a config.edn file from resources and parses it with Aero. (See
    https://github.com/juxt/aero). Two additional reader tags are supported:
@@ -134,5 +134,5 @@
    - keys with a namespace of \"biff.system-properties\" are also merged into
      the system properties."
   []
-  {:biff.core/id    :biff.config/component
+  {:biff.core/id    :biff.config/module
    :biff.core/start start})