+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
-**Breaking**: [2.0.0-rc22](RELEASES/2.0.0-rc22.md) (2026-08-29)
+**Breaking**: [2.0.0-rc23](RELEASES/2.0.0-rc23.md) (2026-08-29)
 
-- biff.core: module-style components
+- biff.core: components are now module lifecycle functions
 - biff.fx: state functions have additional arguments
 - biff.ring: defroute is removed
 - biff.ring: path only takes template strings, not route vectors